Support 64-bit I/O on 32-bit platforms

+++ blink/iovs.h
@@ -15,6 +15,6 @@ struct Iovs {
 void FreeIovs(struct Iovs *);
 void InitIovs(struct Iovs *);
 int AppendIovsReal(struct Machine *, struct Iovs *, i64, u64);
-int AppendIovsGuest(struct Machine *, struct Iovs *, i64, i64);
+int AppendIovsGuest(struct Machine *, struct Iovs *, i64, int);
 
 #endif /* BLINK_IOVS_H_ */
